test(datetime-button): migrate multiple
